election - split code and tests
